+++ Editor/PropertyDrawers/AUEEventPropertyDrawer.cs
@@ -13,61 +13,71 @@ namespace AUE
     [CustomPropertyDrawer(typeof(BaseAUEEvent), useForChildren: true)]
     public class AUEEventPropertyDrawer : PropertyDrawer
     {
-        private const string EventsSPName = "_events";
+        public const string EventsSPName = "_events";
 
-        private ReorderableList _reorderableList = null;
-        private GUIContent _label;
-        private bool _isStateInitialized = false;
+        private class State
+        {
+            public ReorderableList ReorderableList = null;
+            public GUIContent Label;
+            public bool IsStateInitialized = false;
+            public Type[] CachedArgumentTypes;
+        }
+
+        private Dictionary<string, State> _states = new Dictionary<string, State>();
 
         private static readonly MethodInfo DoListHeaderMI = typeof(ReorderableList).GetMethod("DoListHeader", BindingFlags.NonPublic | BindingFlags.Instance);
 
         public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
         {
-            InitReorderableList(property);
-            InitializeState(property);
-            float height = property.isExpanded ? _reorderableList.GetHeight() : EditorGUIUtility.singleLineHeight;
+            State state = CreateOrGetCachedState(property);
+            InitReorderableList(property, state);
+            InitializeState(property, state);
+            float height = property.isExpanded ? state.ReorderableList.GetHeight() : EditorGUIUtility.singleLineHeight;
             height += EditorGUIUtility.standardVerticalSpacing;
             return height;
         }
 
         public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
         {
-            CacheLabel(property, label);
-            InitReorderableList(property);
-            InitializeState(property);
+            State state = CreateOrGetCachedState(property);
+            CacheLabel(property, label, state);
+            InitReorderableList(property, state);
+            InitializeState(property, state);
             if (property.isExpanded)
             {
-                _reorderableList.DoList(position);
+                state.ReorderableList.DoList(position);
             }
             else
             {
-                DrawOnlyHeader(position);
+                DrawOnlyHeader(position, state.ReorderableList);
             }
         }
 
-        private void InitializeState(SerializedProperty property)
+        private void InitializeState(SerializedProperty property, State state)
         {
-            if (_isStateInitialized)
+            if (state.IsStateInitialized)
             {
                 return;
             }
 
             property.isExpanded = (property.FindPropertyRelative(EventsSPName).arraySize > 0);
-            _isStateInitialized = true;
+            state.IsStateInitialized = true;
         }
 
-        private void CacheLabel(SerializedProperty property, GUIContent baseLabel)
+        private void CacheLabel(SerializedProperty property, GUIContent baseLabel, State state)
         {
-            if (_label != null)
+            if (state.Label != null && !HasArgumentsChanged(property, state))
             {
                 return;
             }
 
             var argumentTypesSP = property.FindPropertyRelative(AUEUtils.ArgumentTypesSPName);
             var argumentTypesSB = new StringBuilder();
+            state.CachedArgumentTypes = new Type[argumentTypesSP.arraySize];
             for (int i = 0; i < argumentTypesSP.arraySize; ++i)
             {
                 Type argumentType = SerializableTypeHelper.LoadType(argumentTypesSP.GetArrayElementAtIndex(i));
+                state.CachedArgumentTypes[i] = argumentType;
                 if (argumentType != null)
                 {
                     argumentTypesSB.Append(AUEUtils.MakeHumanDisplayType(argumentType));
@@ -81,23 +91,49 @@ namespace AUE
                     argumentTypesSB.Append(", ");
                 }
             }
-            _label = new GUIContent($"{baseLabel.text}({argumentTypesSB.ToString()})", baseLabel.tooltip);
+            state.Label = new GUIContent($"{baseLabel.text}({argumentTypesSB.ToString()})", baseLabel.tooltip);
         }
 
-        private void DrawOnlyHeader(Rect position)
+        private bool HasArgumentsChanged(SerializedProperty property, State state)
         {
-            DoListHeaderMI.Invoke(_reorderableList, new object[] { position });
+            Type[] cachedArgumentTypes = state.CachedArgumentTypes;
+            if (cachedArgumentTypes == null)
+            {
+                return true;
+            }
+
+            var argumentTypesSP = property.FindPropertyRelative(AUEUtils.ArgumentTypesSPName);
+            if (cachedArgumentTypes.Length != argumentTypesSP.arraySize)
+            {
+                return true;
+            }
+
+            for (int i = 0; i < argumentTypesSP.arraySize; ++i)
+            {
+                Type argumentType = SerializableTypeHelper.LoadType(argumentTypesSP.GetArrayElementAtIndex(i));
+                if (cachedArgumentTypes[i] != argumentType)
+                {
+                    return true;
+                }
+            }
+
+            return false;
         }
 
-        private void InitReorderableList(SerializedProperty property)
+        private void DrawOnlyHeader(Rect position, ReorderableList reorderableList)
         {
-            if (_reorderableList != null)
+            DoListHeaderMI.Invoke(reorderableList, new object[] { position });
+        }
+
+        private void InitReorderableList(SerializedProperty property, State state)
+        {
+            if (state.ReorderableList != null)
             {
                 return;
             }
 
             var eventsSP = property.FindPropertyRelative(EventsSPName);
-            _reorderableList = new ReorderableList(property.serializedObject, eventsSP, draggable: true, displayHeader: true, displayAddButton: true, displayRemoveButton: true)
+            state.ReorderableList = new ReorderableList(property.serializedObject, eventsSP, draggable: true, displayHeader: true, displayAddButton: true, displayRemoveButton: true)
             {
                 onAddCallback = (rol) =>
                 {
@@ -112,20 +148,34 @@ namespace AUE
                     {
                         headerRect.xMin += 10;
                         headerRect.height = 18f;
-                        property.isExpanded = EditorGUI.Foldout(headerRect, property.isExpanded, _label);
+                        property.isExpanded = EditorGUI.Foldout(headerRect, property.isExpanded, state.Label);
                     }
                     EditorGUI.indentLevel = indentLevel;
                 },
                 drawElementCallback = (Rect rect, int index, bool isActive, bool isFocused) =>
                 {
-                    var eventSP = eventsSP.GetArrayElementAtIndex(index);
-                    SyncArgumentTypes(property, eventSP);
-                    EditorGUI.PropertyField(rect, eventSP, eventsSP.isExpanded);
+                    try
+                    {
+                        if (index >= 0 && index < eventsSP.arraySize)
+                        {
+                            var eventSP = eventsSP.GetArrayElementAtIndex(index);
+                            SyncArgumentTypes(property, eventSP);
+                            EditorGUI.PropertyField(rect, eventSP, eventsSP.isExpanded);
+                        }
+                    }
+                    catch (Exception ex)
+                    {
+                        Debug.LogException(ex);
+                    }
                 },
                 elementHeightCallback = (index) =>
                 {
-                    var eventSP = eventsSP.GetArrayElementAtIndex(index);
-                    return EditorGUI.GetPropertyHeight(eventSP, eventsSP.isExpanded);
+                    if (index >= 0 && index < eventsSP.arraySize)
+                    {
+                        var eventSP = eventsSP.GetArrayElementAtIndex(index);
+                        return EditorGUI.GetPropertyHeight(eventSP, eventsSP.isExpanded);
+                    }
+                    return 0.0f;
                 },
             };
         }
@@ -136,9 +186,7 @@ namespace AUE
             aueEventSP.FindPropertyRelative(AUEUtils.MethodNameSPName).stringValue = string.Empty;
             aueEventSP.FindPropertyRelative(AUEUtils.CallStateSPName).enumValueIndex = (int)UnityEventCallState.RuntimeOnly;
 
-            aueEventSP.FindPropertyRelative(AUEUtils.BindingFlagsSPName).intValue = (int)
-            (BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.GetField
-            | BindingFlags.GetProperty | BindingFlags.SetProperty | BindingFlags.SetField);
+            aueEventSP.FindPropertyRelative(AUEUtils.BindingFlagsSPName).intValue = (int)DefaultBindingFlags.AUESimpleMethod;
 
             SyncArgumentTypes(property, aueEventSP);
         }
@@ -154,6 +202,16 @@ namespace AUE
                 var aueMethodArgumentTypeSP = aueMethodArgumentTypesSP.GetArrayElementAtIndex(i);
                 SerializableTypeHelper.CopySerializableType(aueArgumentTypeSP, aueMethodArgumentTypeSP);
             }
+        }
+
+        private State CreateOrGetCachedState(SerializedProperty sp)
+        {
+            if (!_states.TryGetValue(sp.propertyPath, out State state))
+            {
+                state = new State();
+                _states.Add(sp.propertyPath, state);
+            }
+            return state;
         }
     }
 }