+++ Editor/PropertyDrawers/SerializableTypePropertyDrawer.cs
@@ -1,6 +1,5 @@
 using System;
-using System.Collections;
-using System.Collections.Generic;
+using TypeCodebase;
 using UnityEditor;
 using UnityEditor.IMGUI.Controls;
 using UnityEngine;
@@ -18,20 +17,21 @@ namespace AUE
         private bool _hasNewTypeSelected = false;
         private Type _newTypeSelected;
 
-        public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
-        {
-            return EditorGUIUtility.singleLineHeight;
-        }
+        public override float GetPropertyHeight(SerializedProperty property, GUIContent label) => EditorGUIUtility.singleLineHeight;
 
         public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
         {
             InitializeTypeSearchDropdownIFN(property);
             Rect valueRect = EditorGUI.PrefixLabel(position, label);
             Type currentType = SerializableTypeHelper.LoadType(property);
-            if (GUI.Button(valueRect, currentType?.FullName ?? "<undefined>"))
-            {
-                _typeSearchDropdown.Show(valueRect);
-            }
+
+            TypeSelectorGUI.Draw(valueRect, currentType, new TypeSelectorAdvancedDropdown.Settings()
+            { 
+                UsageFlags = TypeCodebase.ETypeUsageFlag.Class | 
+                TypeCodebase.ETypeUsageFlag.Abstract | 
+                TypeCodebase.ETypeUsageFlag.Interface | 
+                TypeCodebase.ETypeUsageFlag.Struct 
+            }, out _hasNewTypeSelected, out _newTypeSelected);
 
             CheckNewTypeSelected(property);
         }
@@ -44,7 +44,7 @@ namespace AUE
             }
 
             var typeFullNameSP = property.FindPropertyRelative(SerializableTypeHelper.SerializedTypeFullNameSPName);
-            typeFullNameSP.stringValue = _newTypeSelected.AssemblyQualifiedName;
+            typeFullNameSP.stringValue = _newTypeSelected?.AssemblyQualifiedName ?? string.Empty;
             _hasNewTypeSelected = false;
         }