+++ Runtime/Core/AUEMethod.cs
@@ -59,7 +59,7 @@ namespace AUE
             _parameterInfos = new AUEMethodParameterInfo[parameters.Length];
             for (int i = 0; i < parameters.Length; ++i)
             {
-                _parameterInfos[i] = new AUEMethodParameterInfo(desc.Parameters[i], parameters[i].ParameterType);
+                _parameterInfos[i] = new AUEMethodParameterInfo(desc.Parameters[i]);
             }
         }
 
@@ -115,9 +115,10 @@ namespace AUE
         {
             OnEachInternalMethod((method) =>
                 {
-                    for (int i = 0; i < ParameterInfos.Length; ++i)
+                    AUEMethodParameterInfo[] parameters = method.ParameterInfos;
+                    for (int i = 0; i < parameters.Length; ++i)
                     {
-                        if (ParameterInfos[i].CustomArgument is T typedCustomArgument)
+                        if (parameters[i].CustomArgument is T typedCustomArgument)
                         {
                             callback(typedCustomArgument);
                         }